Fix mask behaviour when service is enabled
Mask and enabled operations are actually mutually exclusive ones.
So if service is enabled, it can't be masked at the same time.
As a result we need to mask service after it being disabled and unmask
it before enabling.

Currently mask just do not work if service is enabled, so module needs
to be runned twice to apply
{enabled: false, masked: true} on enabld service.
This behaviour should be fixed now.

Fixes: #68680